Update token decimals to comply with ERC20 standard

+++ contracts/GlobalToken.sol
@@ -7,7 +7,7 @@ contract GlobalToken is TimeLockedToken, MintableToken {
     // Standard constants for token readability
     string public constant name = "Global ICO Token";
     string public constant symbol = "GLIF";
-    uint256 public constant decimals = 18;
+    uint8 public constant decimals = 18;
     
     function GlobalToken(uint releaseDate) TimeLockedToken(releaseDate) public {
         totalSupply = 0;
